pyproject.toml
[tool.poetry]
name = "psql-connector"
version = "0.1.4"
description = ""
authors = ["Bastiancito <bastian.gamboa@ug.uchile.cl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.9"
numpy = "^1.26.4"
pandas = "^2.2.2"
sqlalchemy = "^2.0.30"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
